Move fix reports (classes and schemas)
