+++ .gitattributes
@@ -5,7 +5,7 @@
 /tests/Differ/AbstractDifferTestCase.php export-ignore
 /tests/Fixtures/ export-ignore
 /tests/Linter/AbstractLinterTestCase.php export-ignore
-/tests/Report/AbstractReporterTestCase.php export-ignore
+/tests/Console/Report/FixReport/AbstractReporterTestCase.php export-ignore
 /tests/Test/AbstractFixerWithAliasedOptionsTestCase.php export-ignore
 /tests/Test/AbstractTransformerTestCase.php export-ignore
