+++ doc/schemas/fix/junit-10.xsd
