+++ src/Console/Report/FixReport/JsonReporter.php
@@ -10,7 +10,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Report;
+namespace PhpCsFixer\Console\Report\FixReport;
 
 use Symfony\Component\Console\Formatter\OutputFormatter;
