+++ src/Console/Report/FixReport/ReportSummary.php
@@ -10,7 +10,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Report;
+namespace PhpCsFixer\Console\Report\FixReport;
 
 /**
  * @author Dariusz Rumiński <dariusz.ruminski@gmail.com>
