+++ tests/Console/Report/FixReport/GitlabReporterTest.php
@@ -10,15 +10,15 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Tests\Report;
+namespace PhpCsFixer\Tests\Console\Report\FixReport;
 
-use PhpCsFixer\Report\GitlabReporter;
+use PhpCsFixer\Console\Report\FixReport\GitlabReporter;
 
 /**
  * @author Hans-Christian Otto <c.otto@suora.com>
  *
  * @internal
- * @covers \PhpCsFixer\Report\GitlabReporter
+ * @covers \PhpCsFixer\Console\Report\FixReport\GitlabReporter
  */
 final class GitlabReporterTest extends AbstractReporterTestCase
 {
