+++ tests/Console/Report/FixReport/JsonReporterTest.php
@@ -10,9 +10,9 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Tests\Report;
+namespace PhpCsFixer\Tests\Console\Report\FixReport;
 
-use PhpCsFixer\Report\JsonReporter;
+use PhpCsFixer\Console\Report\FixReport\JsonReporter;
 
 /**
  * @author Boris Gorbylev <ekho@ekho.name>
@@ -20,7 +20,7 @@ use PhpCsFixer\Report\JsonReporter;
  *
  * @internal
  *
- * @covers \PhpCsFixer\Report\JsonReporter
+ * @covers \PhpCsFixer\Console\Report\FixReport\JsonReporter
  */
 final class JsonReporterTest extends AbstractReporterTestCase
 {
@@ -153,7 +153,7 @@ JSON;
      */
     private static function assertJsonSchema($json)
     {
-        $jsonPath = __DIR__.'/../../doc/report-schema/schema.json';
+        $jsonPath = __DIR__.'/../../../../doc/schemas/fix/schema.json';
 
         $data = json_decode($json);
