+++ tests/Console/Report/FixReport/ReportSummaryTest.php
@@ -10,9 +10,9 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Tests\Report;
+namespace PhpCsFixer\Tests\Console\Report\FixReport;
 
-use PhpCsFixer\Report\ReportSummary;
+use PhpCsFixer\Console\Report\FixReport\ReportSummary;
 use PhpCsFixer\Tests\TestCase;
 
 /**
@@ -20,7 +20,7 @@ use PhpCsFixer\Tests\TestCase;
  *
  * @internal
  *
- * @covers \PhpCsFixer\Report\ReportSummary
+ * @covers \PhpCsFixer\Console\Report\FixReport\ReportSummary
  */
 final class ReportSummaryTest extends TestCase
 {
