+++ tests/Console/Report/FixReport/TextReporterTest.php
@@ -10,9 +10,9 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Tests\Report;
+namespace PhpCsFixer\Tests\Console\Report\FixReport;
 
-use PhpCsFixer\Report\TextReporter;
+use PhpCsFixer\Console\Report\FixReport\TextReporter;
 
 /**
  * @author Boris Gorbylev <ekho@ekho.name>
@@ -20,7 +20,7 @@ use PhpCsFixer\Report\TextReporter;
  *
  * @internal
  *
- * @covers \PhpCsFixer\Report\TextReporter
+ * @covers \PhpCsFixer\Console\Report\FixReport\TextReporter
  */
 final class TextReporterTest extends AbstractReporterTestCase
 {
